+++ src/sysfs/crawl.rs
@@ -1,0 +1,34 @@
+/// Crawls a directory structure for filenames matching given input
+pub fn crawl(
+    dir: &std::path::Path,
+    filename_regex: &regex::Regex,
+    paths: &mut std::vec::Vec<std::path::PathBuf>,
+) -> std::io::Result<()> {
+    if dir.is_dir() {
+        log::debug!("Checking dir {:?}", dir);
+        for entry in std::fs::read_dir(dir)? {
+            let entry: std::fs::DirEntry = entry?;
+            let path = entry.path();
+            log::debug!("Checking path {:?}", path);
+            // Skip symlinks to avoid infinite loops
+            if path.is_symlink() {
+                continue;
+            }
+
+            // dirs need to be crawled further
+            if path.is_dir() {
+                crawl(&path, filename_regex, paths)?;
+            } else {
+                match path.to_str() {
+                    Some(path_str) => {
+                        if filename_regex.is_match(path_str) {
+                            paths.push(path);
+                        }
+                    }
+                    None => {}
+                }
+            }
+        }
+    }
+    Ok(())
+}
